+++ packages/sdk-core/src/SuperToken.ts
@@ -653,6 +653,28 @@ export class WrapperSuperToken extends SuperToken {
     };
 
     /**
+     * Downgrade `amount` of an ERC20 token to its SuperToken to `to` address.
+     * @param amount The amount to be downgraded.
+     * @param to The destination of the downgraded ERC20 token.
+     * @param overrides ethers overrides object for more control over the transaction sent.
+     * @returns {Operation} An instance of Operation which can be executed.
+     */
+    downgradeTo = ({
+        amount,
+        to,
+        overrides,
+    }: {
+        amount: string;
+        to: string;
+        overrides?: Overrides & { from?: string | Promise<string> };
+    }) => {
+        const txn = this.contract.populateTransaction.downgradeTo(to, amount, {
+            ...overrides,
+        });
+        return new Operation(txn, "UNSUPPORTED");
+    };
+
+    /**
      * Upgrade `amount` SuperToken's.
      * @param amount The amount to be upgraded.
      * @param overrides ethers overrides object for more control over the transaction sent.
@@ -675,8 +697,8 @@ export class WrapperSuperToken extends SuperToken {
     /**
      * Upgrade `amount` of an ERC20 token to its SuperToken to `to` address.
      * @param amount The amount to be upgraded.
-     * @param to The destination of the upgraded native asset super tokens.
-     * @param data Bytes operatorData
+     * @param to The destination of the upgraded wrapper super tokens.
+     * @param data Bytes userData
      * @param overrides ethers overrides object for more control over the transaction sent.
      * @returns {Operation} An instance of Operation which can be executed.
      */